requirements.txt
cffi==1.15.1
cryptography==39.0.1
javaobj-py3==0.4.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
pyjks==20.0.0
pyOpenSSL==23.0.0
twofish==0.3.0
